Begin the grand refactor
