+++ frontend/src/components/GoalForm/__tests__/ObjectiveForm.js
@@ -5,17 +5,11 @@ import {
 } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import fetchMock from 'fetch-mock';
-import selectEvent from 'react-select-event';
 import ObjectiveForm from '../ObjectiveForm';
 import UserContext from '../../../UserContext';
+import { OBJECTIVE_ERROR_MESSAGES } from '../constants';
 
-import {
-  OBJECTIVE_ERROR_MESSAGES,
-} from '../constants';
-
-const [
-  objectiveTextError, objectiveTopicsError,
-] = OBJECTIVE_ERROR_MESSAGES;
+const [objectiveTextError] = OBJECTIVE_ERROR_MESSAGES;
 
 describe('ObjectiveForm', () => {
   const defaultObjective = {
@@ -105,62 +99,10 @@ describe('ObjectiveForm', () => {
 
     renderObjectiveForm(objective, removeObjective, setObjectiveError, setObjective);
 
-    const topics = await screen.findByLabelText(/topics \*/i, { selector: '#topics' });
-    userEvent.click(topics);
-
-    const resourceOne = await screen.findByRole('textbox', { name: 'Resource 1' });
-    userEvent.click(resourceOne);
-
-    expect(setObjectiveError).toHaveBeenCalledWith(index, [<></>, <span className="usa-error-message">{objectiveTopicsError}</span>, <></>, <></>, <></>, <></>]);
-
-    await selectEvent.select(topics, ['Coaching', 'Communication']);
-
-    userEvent.click(topics);
-    userEvent.click(resourceOne);
-    expect(setObjectiveError).toHaveBeenCalledWith(
-      index, [<></>, <></>, <></>, <></>, <></>, <></>,
-      ],
-    );
-
     const objectiveText = await screen.findByRole('textbox', { name: /TTA objective \*/i });
     userEvent.click(objectiveText);
-    userEvent.click(resourceOne);
+    userEvent.tab(); // trigger blur event
 
     expect(setObjectiveError).toHaveBeenCalledWith(index, [<span className="usa-error-message">{objectiveTextError}</span>, <></>, <></>, <></>, <></>, <></>]);
-  });
-
-  it('you can change status', async () => {
-    const removeObjective = jest.fn();
-    const setObjectiveError = jest.fn();
-    const setObjective = jest.fn();
-
-    renderObjectiveForm(
-      { ...defaultObjective, status: 'In Progress' },
-      removeObjective,
-      setObjectiveError,
-      setObjective,
-      'In Progress',
-    );
-
-    const statusSelect = await screen.findByLabelText('Objective status');
-    userEvent.selectOptions(statusSelect, 'Complete');
-    expect(setObjective).toHaveBeenCalledWith({ ...defaultObjective, status: 'Complete' });
-    userEvent.click(statusSelect);
-  });
-
-  it('displays the correct label based on resources from api', async () => {
-    const removeObjective = jest.fn();
-    const setObjectiveError = jest.fn();
-    const setObjective = jest.fn();
-
-    renderObjectiveForm(
-      defaultObjective,
-      removeObjective,
-      setObjectiveError,
-      setObjective,
-    );
-
-    const label = await screen.findByText('Link to TTA resource');
-    expect(label).toBeVisible();
   });
 });
