+++ frontend/src/components/GoalForm/__tests__/index.js
@@ -6,8 +6,6 @@ import {
   screen,
   within,
   waitFor,
-  act,
-  fireEvent,
 } from '@testing-library/react';
 import { REPORT_STATUSES, SCOPE_IDS } from '@ttahub/common';
 import selectEvent from 'react-select-event';
@@ -21,9 +19,7 @@ import { OBJECTIVE_ERROR_MESSAGES } from '../constants';
 import { BEFORE_OBJECTIVES_CREATE_GOAL, BEFORE_OBJECTIVES_SELECT_RECIPIENTS } from '../Form';
 import AppLoadingContext from '../../../AppLoadingContext';
 
-const [
-  objectiveTitleError, objectiveTopicsError,
-] = OBJECTIVE_ERROR_MESSAGES;
+const [objectiveTitleError] = OBJECTIVE_ERROR_MESSAGES;
 
 const topicsFromApi = [
   'Behavioral / Mental Health / Trauma',
@@ -212,17 +208,6 @@ describe('create goal', () => {
     const objectiveText = await screen.findByRole('textbox', { name: /TTA objective \*/i });
     userEvent.type(objectiveText, 'test');
 
-    const topics = await screen.findByLabelText(/topics \*/i, { selector: '#topics' });
-    await selectEvent.select(topics, ['CLASS: Instructional Support']);
-
-    const resourceOne = document.querySelector('#resource-1');
-    userEvent.type(resourceOne, 'https://search.marginalia.nu/');
-
-    const supportType = await screen.findByRole('combobox', { name: /support type/i });
-    act(() => {
-      userEvent.selectOptions(supportType, 'Implementing');
-    });
-
     userEvent.click(save);
 
     expect(fetchMock.called('/api/goals')).toBeTruthy();
@@ -360,17 +345,6 @@ describe('create goal', () => {
     const objectiveText = await screen.findByRole('textbox', { name: /TTA objective \*/i });
     userEvent.type(objectiveText, 'test');
 
-    const topics = await screen.findByLabelText(/topics \*/i, { selector: '#topics' });
-    await selectEvent.select(topics, ['CLASS: Instructional Support']);
-
-    const resourceOne = await screen.findByRole('textbox', { name: 'Resource 1' });
-    userEvent.type(resourceOne, 'https://search.marginalia.nu/');
-
-    const supportType = await screen.findByRole('combobox', { name: /support type/i });
-    act(() => {
-      userEvent.selectOptions(supportType, 'Implementing');
-    });
-
     expect(fetchMock.called('/api/goals')).toBe(false);
     userEvent.click(save);
 
@@ -422,20 +396,6 @@ describe('create goal', () => {
 
     const objectiveText = await screen.findByRole('textbox', { name: /TTA objective \*/i });
     userEvent.type(objectiveText, 'test');
-
-    const topicsText = screen.queryAllByLabelText(/topics \*/i);
-    expect(topicsText.length).toBe(1);
-    const topics = document.querySelector('#topics');
-
-    await selectEvent.select(topics, ['CLASS: Instructional Support']);
-
-    const resourceOne = await screen.findByRole('textbox', { name: 'Resource 1' });
-    userEvent.type(resourceOne, 'https://search.marginalia.nu/');
-
-    const supportType = await screen.findByRole('combobox', { name: /support type/i });
-    act(() => {
-      userEvent.selectOptions(supportType, 'Implementing');
-    });
 
     const save = await screen.findByRole('button', { name: /save and continue/i });
     userEvent.click(save);
@@ -493,20 +453,6 @@ describe('create goal', () => {
     let objectiveText = await screen.findByRole('textbox', { name: /TTA objective \*/i });
     userEvent.type(objectiveText, 'test');
 
-    const topicsText = screen.queryAllByLabelText(/topics \*/i);
-    expect(topicsText.length).toBe(1);
-    let topics = document.querySelector('#topics');
-
-    await selectEvent.select(topics, ['CLASS: Instructional Support']);
-
-    let supportType = await screen.findByRole('combobox', { name: /support type/i });
-    act(() => {
-      userEvent.selectOptions(supportType, 'Implementing');
-    });
-
-    let resourceOne = await screen.findByRole('textbox', { name: 'Resource 1' });
-    userEvent.type(resourceOne, 'https://search.marginalia.nu/');
-
     const cancel = await screen.findByRole('link', { name: 'Cancel' });
     let save = await screen.findByRole('button', { name: /save and continue/i });
     userEvent.click(save);
@@ -545,17 +491,6 @@ describe('create goal', () => {
 
     objectiveText = await screen.findByRole('textbox', { name: /TTA objective \*/i });
     userEvent.type(objectiveText, 'test');
-
-    topics = document.querySelector('#topics');
-    await selectEvent.select(topics, ['CLASS: Instructional Support']);
-
-    supportType = await screen.findByRole('combobox', { name: /support type/i });
-    act(() => {
-      userEvent.selectOptions(supportType, 'Implementing');
-    });
-
-    resourceOne = await screen.findByRole('textbox', { name: 'Resource 1' });
-    userEvent.type(resourceOne, 'https://search.marginalia.nu/');
 
     userEvent.click((await screen.findByRole('button', { name: /save draft/i })));
 
@@ -600,17 +535,6 @@ describe('create goal', () => {
 
     const objectiveText = await screen.findByRole('textbox', { name: /TTA objective \*/i });
     userEvent.type(objectiveText, 'test');
-
-    const topics = await screen.findByLabelText(/topics \*/i, { selector: '#topics' });
-    await selectEvent.select(topics, ['CLASS: Instructional Support']);
-
-    const supportType = await screen.findByRole('combobox', { name: /support type/i });
-    act(() => {
-      userEvent.selectOptions(supportType, 'Implementing');
-    });
-
-    const resourceOne = await screen.findByRole('textbox', { name: 'Resource 1' });
-    userEvent.type(resourceOne, 'https://search.marginalia.nu/');
 
     let save = await screen.findByRole('button', { name: /save and continue/i });
     userEvent.click(save);
@@ -705,18 +629,6 @@ describe('create goal', () => {
 
     userEvent.click(save);
 
-    await screen.findByText(objectiveTopicsError);
-
-    const topics = await screen.findByLabelText(/topics \*/i, { selector: '#topics' });
-    await selectEvent.select(topics, ['Coaching']);
-
-    const supportType = await screen.findByRole('combobox', { name: /support type/i });
-    act(() => {
-      userEvent.selectOptions(supportType, 'Implementing');
-    });
-
-    userEvent.click(save);
-
     await screen.findByText(`Your goal was last saved at ${moment().format('MM/DD/YYYY [at] h:mm a')}`);
 
     expect(cancel).not.toBeVisible();
@@ -724,187 +636,6 @@ describe('create goal', () => {
     const submit = await screen.findByRole('button', { name: /submit goal/i });
     userEvent.click(submit);
     expect(fetchMock.called('/api/goals')).toBe(true);
-  });
-
-  it('can add and validate objective resources', async () => {
-    const recipient = {
-      id: 2,
-      grants: [
-        {
-          id: 2,
-          numberWithProgramTypes: 'Turtle 2',
-          status: 'Active',
-        },
-      ],
-    };
-
-    renderForm(recipient);
-
-    await screen.findByRole('heading', { name: 'Goal summary' });
-    fetchMock.restore();
-    fetchMock.post('/api/goals', postResponse);
-    fetchMock.get('/api/feeds/item?tag=ttahub-topic', `<feed xmlns="http://www.w3.org/2005/Atom" xmlns:dc="http://purl.org/dc/elements/1.1/">
-    <title>Whats New</title>
-    <link rel="alternate" href="https://acf-ohs.atlassian.net/wiki" />
-    <subtitle>Confluence Syndication Feed</subtitle>
-    <id>https://acf-ohs.atlassian.net/wiki</id></feed>`);
-    fetchMock.get('/api/goals/recipient/2/region/1/nudge?name=This%20is%20goal%20text&grantNumbers=undefined', []);
-
-    const goalText = await screen.findByRole('textbox', { name: /Recipient's goal/i });
-    userEvent.type(goalText, 'This is goal text');
-
-    const ed = await screen.findByRole('textbox', { name: /anticipated close date \(mm\/dd\/yyyy\)/i });
-    userEvent.type(ed, '08/15/2023');
-
-    let newObjective = await screen.findByRole('button', { name: 'Add new objective' });
-    userEvent.click(newObjective);
-
-    const objectiveText = await screen.findByRole('textbox', { name: /TTA objective \*/i });
-    userEvent.type(objectiveText, 'This is objective text');
-
-    const topics = await screen.findByLabelText(/topics \*/i, { selector: '#topics' });
-    await selectEvent.select(topics, ['Coaching']);
-
-    const supportType = await screen.findByRole('combobox', { name: /support type/i });
-    act(() => {
-      userEvent.selectOptions(supportType, 'Implementing');
-    });
-
-    const resourceOne = await screen.findByRole('textbox', { name: 'Resource 1' });
-    userEvent.type(resourceOne, 'garrgeler');
-
-    const save = await screen.findByRole('button', { name: /save and continue/i });
-    userEvent.click(save);
-
-    await screen.findByText('Enter one resource per field. Valid resource links must start with http:// or https://');
-
-    userEvent.clear(resourceOne);
-    userEvent.type(resourceOne, 'https://search.marginalia.nu/');
-
-    let addNewResource = await screen.findByRole('button', { name: 'Add new resource' });
-    userEvent.click(addNewResource);
-
-    const resourceTwo = await screen.findByRole('textbox', { name: 'Resource 2' });
-    userEvent.type(resourceTwo, 'https://search.marginalia.nu/https://search.marginalia.nu/https://search.marginalia.nu/');
-
-    const saveDraft = await screen.findByRole('button', { name: /save draft/i });
-    userEvent.click(saveDraft);
-
-    await screen.findByText('Enter one resource per field. Valid resource links must start with http:// or https://');
-
-    userEvent.clear(resourceTwo);
-    userEvent.type(resourceTwo, 'https://search.marginalia.nu/');
-
-    addNewResource = await screen.findByRole('button', { name: 'Add new resource' });
-    userEvent.click(addNewResource);
-
-    const resourceThree = await screen.findByRole('textbox', { name: 'Resource 3' });
-    userEvent.type(resourceThree, 'NOT A LINK NOT A LINK HAHAHAHA');
-
-    userEvent.click(save);
-
-    await screen.findByText('Enter one resource per field. Valid resource links must start with http:// or https://');
-
-    addNewResource = await screen.findByRole('button', { name: 'Add new resource' });
-    userEvent.click(addNewResource);
-
-    const removeResource = await screen.findByRole('button', { name: /remove resource 3/i });
-    userEvent.click(removeResource);
-
-    newObjective = await screen.findByRole('button', { name: 'Add new objective' });
-    userEvent.click(newObjective);
-
-    const removeObjective = await screen.findByRole('button', { name: 'Remove objective 2' });
-    userEvent.click(removeObjective);
-
-    userEvent.click(save);
-
-    await screen.findByText(`Your goal was last saved at ${moment().format('MM/DD/YYYY [at] h:mm a')}`);
-
-    const submit = await screen.findByRole('button', { name: /submit goal/i });
-    userEvent.click(submit);
-    expect(fetchMock.called()).toBe(true);
-  });
-
-  it('can add objective files', async () => {
-    const recipient = {
-      id: 2,
-      grants: [
-        {
-          id: 2,
-          numberWithProgramTypes: 'Turtle 2',
-          status: 'Active',
-        },
-      ],
-    };
-
-    renderForm(recipient);
-
-    await screen.findByRole('heading', { name: 'Goal summary' });
-    fetchMock.restore();
-    fetchMock.get('/api/feeds/item?tag=ttahub-topic', `<feed xmlns="http://www.w3.org/2005/Atom" xmlns:dc="http://purl.org/dc/elements/1.1/">
-    <title>Whats New</title>
-    <link rel="alternate" href="https://acf-ohs.atlassian.net/wiki" />
-    <subtitle>Confluence Syndication Feed</subtitle>
-    <id>https://acf-ohs.atlassian.net/wiki</id></feed>`);
-    fetchMock.post('/api/goals', postResponse);
-    fetchMock.get('/api/goals/recipient/2/region/1/nudge?name=This%20is%20goal%20text&grantNumbers=undefined', []);
-
-    const goalText = await screen.findByRole('textbox', { name: /Recipient's goal/i });
-    userEvent.type(goalText, 'This is goal text');
-
-    const ed = await screen.findByRole('textbox', { name: /anticipated close date \(mm\/dd\/yyyy\)/i });
-    userEvent.type(ed, '08/15/2023');
-
-    const newObjective = await screen.findByRole('button', { name: 'Add new objective' });
-    userEvent.click(newObjective);
-
-    expect(document.querySelectorAll('ttahub-objective-files').length).toBe(0);
-
-    const objectiveText = await screen.findByRole('textbox', { name: /TTA objective \*/i });
-    userEvent.type(objectiveText, 'This is objective text');
-
-    const saveDraft = await screen.findByRole('button', { name: /save draft/i });
-    await waitFor(() => userEvent.click(saveDraft));
-
-    const fieldset = document.querySelector('.ttahub-objective-files');
-
-    const yes = await within(fieldset).findByRole('radio', { name: 'Yes' });
-    const no = await within(fieldset).findByRole('radio', { name: 'No' });
-
-    expect(no.checked).toBe(true);
-    act(() => userEvent.click(yes));
-    expect(yes.checked).toBe(true);
-
-    await screen.findByText('Attach any non-link resources');
-
-    const dispatchEvt = (node, type, data) => {
-      const event = new Event(type, { bubbles: true });
-      Object.assign(event, data);
-      fireEvent(node, event);
-    };
-
-    const mockData = (files) => ({
-      dataTransfer: {
-        files,
-        items: files.map((file) => ({
-          kind: 'file',
-          type: file.type,
-          getAsFile: () => file,
-        })),
-        types: ['Files'],
-      },
-    });
-
-    const file = (name, id, status = 'Uploaded') => ({
-      originalFileName: name, id, fileSize: 2000, status, lastModified: 123456,
-    });
-
-    const data = mockData([file('file.csv', 1)]);
-
-    const dropzone = await screen.findByRole('button', { name: 'Select and upload' });
-    act(() => dispatchEvt(dropzone, 'drop', data));
-    await waitFor(() => fetchMock.called('/api/files/objectives'));
   });
 
   it('fetches and prepopulates goal data given an appropriate ID', async () => {
@@ -929,8 +660,6 @@ describe('create goal', () => {
           id: 1238474,
           title: 'This is an objective',
           status: 'Not Started',
-          resources: [],
-          topics: [topicsFromApi[0]],
         },
       ],
     }]);
@@ -945,51 +674,6 @@ describe('create goal', () => {
 
     const endDate = await screen.findByRole('textbox', { name: /anticipated close date/i });
     expect(endDate.value).toBe('10/08/2021');
-  });
-
-  it('objective can be suspended', async () => {
-    fetchMock.get('/api/recipient/1/goals?goalIds=', [{
-      name: 'This is a goal name',
-      status: 'In Progress',
-      endDate: '10/08/2021',
-      goalNumbers: ['G-12389'],
-      isRttapa: null,
-      prompts: [],
-      sources: [],
-      grants: [{
-        id: 1,
-        number: '1',
-        programs: [{
-          programType: 'EHS',
-        }],
-        status: 'Active',
-      }],
-      objectives: [
-        {
-          id: 1238474,
-          title: 'This is an objective',
-          status: 'Not Started',
-          resources: [],
-          topics: [topicsFromApi[0]],
-        },
-      ],
-    }]);
-
-    renderForm(defaultRecipient, '12389');
-    const objectiveStatus = await screen.findByRole('combobox', { name: /objective status/i });
-
-    userEvent.selectOptions(objectiveStatus, 'Suspended');
-    const recipientRequestReason = await screen.findByLabelText(/Recipient request/i);
-    userEvent.click(recipientRequestReason);
-
-    const context = await screen.findByLabelText(/Additional context/i);
-    userEvent.type(context, 'This is the context');
-
-    userEvent.click(await screen.findByText(/Submit/i));
-
-    expect(await screen.findByLabelText(/objective status/i)).toBeVisible();
-    expect(await screen.findByText(/reason suspended/i)).toBeVisible();
-    expect(await screen.findByText(/recipient request - this is the context/i)).toBeVisible();
   });
 
   it('draft goals don\'t show status dropdowns', async () => {
@@ -1013,8 +697,6 @@ describe('create goal', () => {
           id: 1238474,
           title: 'This is an objective',
           status: 'Not Started',
-          resources: [],
-          topics: [topicsFromApi[0]],
         },
       ],
     }]);
@@ -1055,8 +737,6 @@ describe('create goal', () => {
           id: 1238474,
           title: 'This is an objective',
           status: 'Not Started',
-          resources: [],
-          topics: [topicsFromApi[0]],
           activityReports: [
             {
               status: REPORT_STATUSES.SUBMITTED,
